analysis: move all llm related to components/common/llmAnalysis
