+++ dashboard/new-dashboard/src/components/common/sideBar/ReportMetricDialog.vue
@@ -151,7 +151,7 @@ import { computed, ref, watch } from "vue"
 import { useToast } from "primevue/usetoast"
 import { Accident, AccidentKind, AccidentsConfigurator } from "../../../configurators/accidents/AccidentsConfigurator"
 import { LlmAnalysesConfigurator } from "../../../configurators/llmAnalyses/LlmAnalysesConfigurator"
-import { startLlmAnalysisWithToast } from "./LlmAnalysisUtils"
+import { startLlmAnalysisWithToast } from "../llmAnalysis/LlmAnalysisUtils"
 import { InfoData } from "./InfoSidebar"
 import { generateDefaultReason } from "./AccidentUtils"
 import { detectPossibleMisclick } from "./MisclickHeuristic"
